+++ apps/backend/prisma/migrations/20240804210316_team_invitation/migration.sql
@@ -1,0 +1,5 @@
+-- AlterEnum
+ALTER TYPE "EmailTemplateType" ADD VALUE 'TEAM_INVITATION';
+
+-- AlterEnum
+ALTER TYPE "VerificationCodeType" ADD VALUE 'TEAM_INVITATION';