+++ apps/e2e/tests/backend/endpoints/api/v1/auth/password/sign-up.test.ts
@@ -26,7 +26,22 @@ it("should sign up new users", async ({ expect }) => {
       },
     ]
   `);
-  await Auth.expectToBeSignedIn();
+  const response = await Auth.expectToBeSignedIn();
+  expect(response.body.auth_methods).toMatchInlineSnapshot(`
+    [
+      {
+        "identifier": "<stripped UUID>@stack-generated.example.com",
+        "type": "password",
+      },
+      {
+        "contact_channel": {
+          "email": "<stripped UUID>@stack-generated.example.com",
+          "type": "email",
+        },
+        "type": "otp",
+      },
+    ]
+  `);
 });
 
 it("should not allow signing up with an e-mail that already exists", async ({ expect }) => {