+++ examples/cjs-test/CHANGELOG.md
@@ -1,5 +1,12 @@
 # cjs-test
 
+## 2.5.17
+
+### Patch Changes
+
+- Bugfixes
+  - @stackframe/stack@2.5.17
+
 ## 2.5.16
 
 ### Patch Changes